requirements.txt
fastapi
uvicorn
numpy
joblib
requests
shapely
pillow
mercantile